Make Provider and hooks directly available from package
